Refactored changeIntent() logic to be clearer and fixed a minor bug
When changing a predicate (a condition or effect in a volition or trigger rule), if setting its value to 0 or false, we would treat it as not having a previous value, and would therefore overwrite its value to the default value (as specified in the schema.json). This commit fixes this behavior.